requirements.txt
Flask==2.3.2
Pillow==9.5.0
easyocr==1.7.1
numpy==1.24.4
opencv-python-headless==4.7.0.72
